+++ bugzoo/__init__.py
@@ -1,9 +1,9 @@
-from bugzoo.version import __version__
-from bugzoo.client import Client
-from bugzoo.manager import BugZoo
-from bugzoo.core import Language, Patch
-from bugzoo.core.container import Container
-from bugzoo.core.bug import Bug
+from .version import __version__
+from .client import Client
+from .manager import BugZoo
+from .core import Language, Patch
+from .core.container import Container
+from .core.bug import Bug
 
 import logging
 logging.getLogger(__name__).addHandler(logging.NullHandler())